01-basics/consumer/requirements.txt
pulsar-client==2.4.2